+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 
 ## Unreleased
 ### Changes
-- (None)
+- Add Window::show_with_args() and show_with_env_args() for FLTK specific command-line arguments.
+- Add some missing asserts.
 
 ## [0.7.20] - 2020-07-23
 ### Changes